Flesh out logging
